Fixing build error: call of overloaded ‘sleep(int)’ is ambiguous
I don't know C++, but the build of fictrac was failing for me with the error above. I figured the function "sleep" now exist among some of the includes of the fictrac and is creating a conflict. Changing it here, as well as in timing.h and CVSource.cpp solved the issue for me